virtual keyboard works

+++ test.rb
@@ -1,9 +1,13 @@
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), 'lib'))
+require 'bundler/setup'
 require 'uinput/keyboard'
 
 begin
-  keyboard = Uinput::Keyboard.create
-  keyboard.combo('ctrl+d')
+  keyboard = UInput::Keyboard.create
+  loop do
+    keyboard.tap('d')
+    sleep 1
+  end
 ensure
-  keyboard.destroy
+  keyboard.destroy if keyboard
 end
